deprecated/requirements.txt
Flask
numpy
Pillow
toml
gunicorn
